Fix for new logs/ directory structure
Update mcwho to work with the new log structure in recent versions of
Minecraft. Logs are now zipped at the end of the day and renamed to have
a date stamp. With this update mcwho will read the whole set of logs at
start, and will watch latest.log, reading it whenever it's updated.